+++ src/components/pages/HomePage/DetailsSection.js
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { Divider } from 'antd';
-import { BridgesContext } from '../../state/bridgesContext';
+import { BridgesContext } from '../../../state/bridgesContext';
 
 const Details = ({ props }) => {
   const { bridgeData } = useContext(BridgesContext);
